refactor: moved withRetry to retry.ts
