+++ src/tokens/aoTokens/sync.ts
@@ -17,6 +17,7 @@ import {
   Id,
   Owner
 } from "./ao";
+import { withRetry } from "~utils/retry";
 
 /** Tokens storage name */
 const AO_TOKENS = "ao_tokens";
@@ -35,44 +36,6 @@ const gateway = {
   port: 443,
   protocol: "https"
 };
-
-/**
- * Generic retry function for any async operation.
- * @param fn - The async function to be retried.
- * @param maxRetries - Maximum retry attempts.
- * @param retryDelay - Delay between retries in milliseconds.
- * @returns A promise of the type that the async function returns.
- */
-async function withRetry<T>(
-  fn: () => Promise<T>,
-  maxRetries: number = 3,
-  retryDelay: number = 100
-): Promise<T> {
-  let lastError: any;
-
-  const delay = (ms: number) =>
-    new Promise((resolve) => setTimeout(resolve, ms));
-
-  for (let attempt = 1; attempt <= maxRetries; attempt++) {
-    try {
-      return await fn();
-    } catch (error) {
-      lastError = error;
-      if (attempt < maxRetries) {
-        const waitTime = Math.pow(2, attempt - 1) * retryDelay;
-        console.log(`Attempt ${attempt} failed, retrying in ${waitTime}ms...`);
-        await delay(waitTime);
-      } else {
-        console.error(
-          `All ${maxRetries} attempts failed. Last error:`,
-          lastError
-        );
-      }
-    }
-  }
-
-  throw lastError;
-}
 
 async function getTokenInfo(id: string): Promise<TokenInfo> {
   const body = {
