+++ src/utils/retry.ts
@@ -68,3 +68,41 @@ export async function retryWithDelayAndTimeout<T>(
   // Final fallback error
   throw new Error("Max attempts reached without success.");
 }
+
+/**
+ * Generic retry function for any async operation.
+ * @param fn - The async function to be retried.
+ * @param maxRetries - Maximum retry attempts.
+ * @param retryDelay - Delay between retries in milliseconds.
+ * @returns A promise of the type that the async function returns.
+ */
+export async function withRetry<T>(
+  fn: () => Promise<T>,
+  maxRetries: number = 3,
+  retryDelay: number = 100
+): Promise<T> {
+  let lastError: any;
+
+  const delay = (ms: number) =>
+    new Promise((resolve) => setTimeout(resolve, ms));
+
+  for (let attempt = 1; attempt <= maxRetries; attempt++) {
+    try {
+      return await fn();
+    } catch (error) {
+      lastError = error;
+      if (attempt < maxRetries) {
+        const waitTime = Math.pow(2, attempt - 1) * retryDelay;
+        console.log(`Attempt ${attempt} failed, retrying in ${waitTime}ms...`);
+        await delay(waitTime);
+      } else {
+        console.error(
+          `All ${maxRetries} attempts failed. Last error:`,
+          lastError
+        );
+      }
+    }
+  }
+
+  throw lastError;
+}
